# app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Instagram Message Router - CRM Integration API
  description: |
    API for integrating a CRM system with Instagram messaging.
    
    ## Implementation Status
    
    **Core MVP Purpose:** Enable CRM chat window for Instagram customer communication
    
    **✅ Fully Implemented (Outbound Flow):**
    - `/health` - Health check endpoint
    - `/api/v1/messages/send` - Send messages to Instagram users (CRM → Instagram)
    - `/api/v1/messages/{message_id}/status` - Get message delivery status
    - `/api/v1/accounts` (POST) - Create account configuration
    - `/webhooks/instagram` - Instagram webhook receiver (Instagram → Router)
    
    **❌ CRITICAL MISSING (Inbound Flow):**
    - Webhook forwarding to CRM (Router → CRM) - **BLOCKS CRM CHAT WINDOW**
    - Without this, CRM cannot receive customer messages
    
    **🚧 Partially Implemented:**
    - `/api/v1/accounts/{account_id}` (GET) - Documented but not implemented
    - `/api/v1/accounts/{account_id}` (PUT) - Documented but not implemented
    
    **📋 Planned (Priority 2+):**
    - Webhook retry logic and DLQ
    - Delivery status webhooks
    - Advanced delivery tracking
    
    ---
    
    ## Authentication
    
    All endpoints require Bearer token authentication via the `Authorization` header.
    
    **Development Mode:** Any Bearer token is accepted (stub authentication)
    **Production Mode:** Real API key validation required
    
    Example:
    ```
    Authorization: Bearer your_api_key_here
    ```
    
    ---
    
    ## Local Testing
    
    **For API testing (send/receive messages):**
    - Use `http://localhost:8000` - Direct connection, faster, no ngrok overhead
    
    **For Instagram webhook testing (receive messages from Instagram):**
    - Use ngrok URL (e.g., `https://xxx.ngrok-free.dev`) - Required for Instagram to reach your local server
    - Configure webhook URL in Facebook Developer Console with your ngrok URL
    
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000
    description: Local development (API testing)
  - url: https://your-ngrok-url.ngrok-free.dev
    description: Local development (Instagram webhooks)
  - url: https://api.yourdomain.com
    description: Production

security:
  - BearerAuth: []

paths:
  # ============================================
  # OUTBOUND MESSAGING (CRM → Customer)
  # ============================================
  
  /api/v1/messages/send:
    post:
      summary: Send message to Instagram user
      description: |
        ✅ **IMPLEMENTED**
        
        Send a message from your Instagram business account to a customer.
        Returns 202 Accepted with message_id. Delivery happens synchronously (MVP).
        
        **Note:** Current implementation sends immediately (not async queue).
        Async delivery with retries planned for Priority 2.
        
        **Idempotency:** Use idempotency_key to prevent duplicate sends.
      operationId: sendMessage
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            examples:
              simple:
                summary: Simple text message
                value:
                  account_id: "ig_account_123"
                  recipient_id: "1234567890"
                  message: "Thanks for your order! It will ship tomorrow."
                  idempotency_key: "order_456_confirmation"
      responses:
        '202':
          description: Message accepted for delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
              example:
                message_id: "msg_abc123"
                status: "pending"
                created_at: "2025-11-06T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/messages/{message_id}/status:
    get:
      summary: Get message delivery status
      description: |
        ✅ **IMPLEMENTED**
        
        Query the current delivery status of a sent message.
        Returns detailed status including Instagram message ID and error details if failed.
      operationId: getMessageStatus
      tags:
        - Messages
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: string
          example: "msg_abc123"
      responses:
        '200':
          description: Message status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageStatusResponse'
              example:
                message_id: "msg_abc123"
                status: "delivered"
                account_id: "ig_account_123"
                recipient_id: "1234567890"
                instagram_message_id: "ig_mid_xyz789"
                created_at: "2025-11-06T10:30:00Z"
                sent_at: "2025-11-06T10:30:02Z"
                delivered_at: "2025-11-06T10:30:05Z"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # ACCOUNT MANAGEMENT
  # ============================================
  
  /api/v1/accounts:
    post:
      summary: Create Instagram account configuration
      description: |
        ✅ **IMPLEMENTED**
        
        Register an Instagram business account with the router.
        Returns account_id for use in other API calls.
        
        **Note:** Credentials are base64-encoded (NOT encrypted) in MVP.
        Real encryption planned for Priority 2.
      operationId: createAccount
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
            example:
              instagram_account_id: "17841400000000000"
              username: "myshop"
              access_token: "EAABwzLixnjYBO..."
              crm_webhook_url: "https://crm.myshop.com/webhooks/instagram"
              webhook_secret: "shared_secret_xyz"
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/accounts/{account_id}:
    get:
      summary: Get account details
      description: |
        🚧 **NOT IMPLEMENTED**
        
        Retrieve account configuration (credentials not included).
        This endpoint is documented but not yet implemented.
      operationId: getAccount
      tags:
        - Accounts
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
          example: "ig_account_123"
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    put:
      summary: Update account configuration
      description: |
        🚧 **NOT IMPLEMENTED**
        
        Update webhook URL, credentials, or other settings.
        This endpoint is documented but not yet implemented.
      operationId: updateAccount
      tags:
        - Accounts
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountRequest'
      responses:
        '200':
          description: Account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # HEALTH CHECK
  # ============================================
  
  /health:
    get:
      summary: Health check
      description: |
        ✅ **IMPLEMENTED**
        
        Check if the service is running.
        Returns status and timestamp.
      operationId: healthCheck
      tags:
        - System
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  timestamp:
                    type: string
                    format: date-time

# ============================================
# WEBHOOK PAYLOADS (CRM receives these)
# ============================================

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        API key for authentication. Format: `Bearer router_live_abc123...`
        
        Get your API key from the admin dashboard.

  schemas:
    # ============================================
    # REQUEST SCHEMAS
    # ============================================
    
    SendMessageRequest:
      type: object
      required:
        - account_id
        - recipient_id
        - message
        - idempotency_key
      properties:
        account_id:
          type: string
          description: Instagram account ID to send from
          example: "ig_account_123"
        recipient_id:
          type: string
          description: Instagram PSID of the recipient
          example: "1234567890"
        message:
          type: string
          description: Message text to send
          minLength: 1
          maxLength: 1000
          example: "Thanks for your order!"
        idempotency_key:
          type: string
          description: Unique key to prevent duplicate sends
          example: "order_456_confirmation"
        metadata:
          type: object
          description: Optional CRM-specific metadata
          additionalProperties: true
          example:
            order_id: "456"
            customer_id: "789"

    CreateAccountRequest:
      # Note: app_secret removed from MVP - not needed for account storage
      type: object
      required:
        - instagram_account_id
        - username
        - access_token
        - crm_webhook_url
        - webhook_secret
      properties:
        instagram_account_id:
          type: string
          description: Instagram business account ID
          example: "17841400000000000"
        username:
          type: string
          description: Instagram username
          example: "myshop"
        access_token:
          type: string
          description: Instagram Page Access Token
          example: "EAABwzLixnjYBO..."
        crm_webhook_url:
          type: string
          format: uri
          description: URL where router sends webhooks
          example: "https://crm.myshop.com/webhooks/instagram"
        webhook_secret:
          type: string
          description: Shared secret for webhook signatures
          example: "shared_secret_xyz"

    UpdateAccountRequest:
      type: object
      properties:
        crm_webhook_url:
          type: string
          format: uri
        webhook_secret:
          type: string
        access_token:
          type: string
        status:
          type: string
          enum: [active, inactive]

    # ============================================
    # RESPONSE SCHEMAS
    # ============================================
    
    SendMessageResponse:
      type: object
      properties:
        message_id:
          type: string
          description: Message Router's message ID
          example: "msg_abc123"
        status:
          type: string
          enum: [pending, accepted, failed]
          example: "pending"
        created_at:
          type: string
          format: date-time
          example: "2025-11-06T10:30:00Z"

    MessageStatusResponse:
      type: object
      properties:
        message_id:
          type: string
          example: "msg_abc123"
        status:
          type: string
          enum: [pending, sent, delivered, read, failed]
          example: "delivered"
        account_id:
          type: string
          example: "ig_account_123"
        recipient_id:
          type: string
          example: "1234567890"
        instagram_message_id:
          type: string
          nullable: true
          example: "ig_mid_xyz789"
        created_at:
          type: string
          format: date-time
        sent_at:
          type: string
          format: date-time
          nullable: true
        delivered_at:
          type: string
          format: date-time
          nullable: true
        read_at:
          type: string
          format: date-time
          nullable: true
        error:
          $ref: '#/components/schemas/ErrorDetail'
          nullable: true

    AccountResponse:
      # Note: status field removed from MVP - account status tracking deferred to Priority 2
      type: object
      properties:
        account_id:
          type: string
          example: "ig_account_123"
        instagram_account_id:
          type: string
          example: "17841400000000000"
        username:
          type: string
          example: "myshop"
        crm_webhook_url:
          type: string
          example: "https://crm.myshop.com/webhooks/instagram"
        created_at:
          type: string
          format: date-time

    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          example: "rate_limit"
        message:
          type: string
          example: "Instagram API rate limit exceeded"
        retryable:
          type: boolean
          example: true

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "invalid_request"
        message:
          type: string
          example: "The recipient_id is required"
        details:
          type: object
          additionalProperties: true
        correlation_id:
          type: string
          example: "req_abc123"
        timestamp:
          type: string
          format: date-time

    # ============================================
    # WEBHOOK PAYLOADS (sent to CRM)
    # ============================================
    
    InboundMessageWebhook:
      type: object
      description: |
        Webhook sent when a customer sends a message to your Instagram account.
        
        **Security:** Validate the `X-Hub-Signature-256` header using HMAC-SHA256:
        ```python
        import hmac, hashlib
        signature = hmac.new(webhook_secret.encode(), body, hashlib.sha256).hexdigest()
        is_valid = hmac.compare_digest(f"sha256={signature}", header_value)
        ```
      properties:
        event:
          type: string
          enum: [message.received]
          example: "message.received"
        message_id:
          type: string
          description: Message Router's message ID
          example: "msg_inbound_xyz"
        account_id:
          type: string
          example: "ig_account_123"
        sender_id:
          type: string
          description: Instagram PSID of sender
          example: "1234567890"
        sender_username:
          type: string
          nullable: true
          example: "customer_john"
        message:
          type: string
          example: "When will my order ship?"
        message_type:
          type: string
          enum: [text, image, video, story_reply]
          example: "text"
        timestamp:
          type: string
          format: date-time
          example: "2025-11-06T10:30:00Z"
        instagram_message_id:
          type: string
          example: "ig_mid_abc123"
        conversation_id:
          type: string
          example: "conv_xyz789"

    DeliveryStatusWebhook:
      type: object
      description: |
        Webhook sent when message delivery status changes.
        
        **Events:**
        - message.sent: Successfully sent to Instagram API
        - message.delivered: Instagram confirmed delivery
        - message.read: Customer read the message
        - message.failed: Permanent delivery failure
      properties:
        event:
          type: string
          enum: [message.sent, message.delivered, message.read, message.failed]
          example: "message.delivered"
        message_id:
          type: string
          description: Original message_id from send request
          example: "msg_abc123"
        account_id:
          type: string
          example: "ig_account_123"
        recipient_id:
          type: string
          example: "1234567890"
        status:
          type: string
          example: "delivered"
        timestamp:
          type: string
          format: date-time
        instagram_message_id:
          type: string
          nullable: true
          example: "ig_mid_xyz789"
        error:
          $ref: '#/components/schemas/ErrorDetail'
          nullable: true

  # ============================================
  # REUSABLE RESPONSES
  # ============================================
  
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "invalid_request"
            message: "The recipient_id is required"
            correlation_id: "req_abc123"
            timestamp: "2025-11-06T10:30:00Z"

    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
            message: "Invalid API key"
            correlation_id: "req_abc123"
            timestamp: "2025-11-06T10:30:00Z"

    Forbidden:
      description: No access to resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "forbidden"
            message: "API key does not have access to this account"
            correlation_id: "req_abc123"
            timestamp: "2025-11-06T10:30:00Z"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "not_found"
            message: "Message not found"
            correlation_id: "req_abc123"
            timestamp: "2025-11-06T10:30:00Z"

tags:
  - name: Messages
    description: Send messages and check delivery status
  - name: Accounts
    description: Manage Instagram account configurations
  - name: System
    description: Health checks and system status